refactor: simplified import
