+++ lib/models/models.dart
@@ -1,0 +1,3 @@
+export './product.dart';
+export './store.dart';
+export './user.dart';
