+++ lib/resources/repositories/storeRepository.dart
@@ -1,12 +1,9 @@
 import 'dart:async';
 
-import 'package:Eliverd/models/user.dart';
-import 'package:Eliverd/resources/accountProvider.dart';
 import 'package:meta/meta.dart';
 
-import 'package:Eliverd/resources/storeProvider.dart';
-import 'package:Eliverd/models/product.dart';
-import 'package:Eliverd/models/store.dart';
+import 'package:Eliverd/resources/providers/providers.dart';
+import 'package:Eliverd/models/models.dart';
 
 class StoreRepository {
   final StoreAPIClient storeAPIClient;
@@ -35,28 +32,5 @@ class StoreRepository {
   Future<void> removeStock(String storeId, String productId) async {
     // TO-DO: 구현 이후 수정
     await storeAPIClient.removeStock(storeId, productId);
-  }
-}
-
-class AccountRepository {
-  final AccountAPIClient accountAPIClient;
-
-  AccountRepository({@required this.accountAPIClient})
-      : assert(accountAPIClient != null);
-
-  Future<User> signUpUser(Map<String, dynamic> jsonifiedUser) async {
-    final user = accountAPIClient.signUpUser(jsonifiedUser);
-
-    return user;
-  }
-
-  Future<Session> createSession(String userId, String password) async {
-    final session = accountAPIClient.createSession(userId, password);
-
-    return session;
-  }
-
-  Future<void> deleteSession(String session) async {
-    await accountAPIClient.deleteSession(session);
   }
 }
