+++ lib/ui/pages/home.dart
@@ -1,12 +1,12 @@
 import 'package:Eliverd/common/color.dart';
 import 'package:Eliverd/common/string.dart';
-import 'package:Eliverd/resources/storeProvider.dart';
+import 'package:Eliverd/resources/providers/storeProvider.dart';
 import 'package:flutter/material.dart';
 
 import 'package:Eliverd/ui/widgets/header.dart';
 import 'package:Eliverd/ui/widgets/product.dart';
 
-import 'package:Eliverd/resources/repository.dart';
+import 'package:Eliverd/resources/repositories/repositories.dart';
 import 'package:http/http.dart' as http;
 
 import 'add_product.dart';
